requirements.txt
flask==3.0.3
alpaca-py==0.26.0
gspread==6.1.2
google-auth==2.29.0
python-dotenv==1.0.1
requests==2.31.0
